+++ local.tf
@@ -16,9 +16,9 @@ locals {
   # az aks get-versions --location westeurope --output table
   # https://releases.aks.azure.com/webpage/index.html
   cluster_full_versions = merge({
-    "1.24" = "1.24.6"
+    "1.24" = "1.24.9"
     "1.23" = "1.23.12"
-  }, var.experimental.v1_25 ? { "1.25" = "1.25.2" } : {})
+  }, var.experimental.v1_25 ? { "1.25" = "1.25.5" } : {})
 
   availability_zones = [1, 2, 3]